course_project/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
Pillow==9.5.0
pony==0.7.16
psycopg2==2.9.6
requests==2.28.1
urllib3==1.26.12
vk-api==11.9.9
